Remove port number from package.json
